requirements-dev.txt
# The following are only required for Cement development
pytest
pytest-cov
coverage
sphinx
pep8
autopep8
flake8
guzzle_sphinx_theme
recommonmark
sphinxcontrib-napoleon

# Required for optional extensions
pystache
pyYaml
configobj
pylibmc
redis
colorlog
tabulate
jinja2
watchdog
pybars3
